Decoupled mongo setup from environment config
