+++ monkey/monkey_island.py
@@ -3,7 +3,6 @@ import monkey_island.setup.gevent_setup  # noqa: F401 isort:skip
 
 import json
 
-import monkey_island.cc.environment.environment_singleton as env_singleton
 from monkey_island.cc.arg_parser import parse_cli_args
 from monkey_island.cc.server_utils.island_logger import setup_logging
 from monkey_island.setup.config_setup import setup_config_by_cmd_arg, setup_default_config
@@ -32,8 +31,7 @@ if "__main__" == __name__:
     # We need to initialize environment singleton before importing main,
     # because main imports modules from monkey_island/cc/models and models need a connection to the
     # mongodb. Mongodb connection parameters are initialized in environment singleton.
-    env_singleton.initialize_from_file(server_config_path)
 
-    from monkey_island.cc.main import main  # noqa: E402
+    from monkey_island.cc.server_setup import setup_island  # noqa: E402
 
-    main(island_args.setup_only, config)
+    setup_island(island_args.setup_only, config, server_config_path)
