+++ monkey/monkey_island/cc/setup/mongo/database_initializer.py
