+++ monkey/monkey_island/cc/setup/mongo/mongo_process_runner.py
